convert to new csproj and use BenchmarkDotNet
