+++ .gitignore
@@ -194,3 +194,6 @@ FakesAssemblies/
 
 # Visual Studio 6 workspace options file
 *.opt
+
+.idea
+BenchmarkDotNet.Artifacts*
