--- JsEnginePerformanceComparison/packages.config
@@ -1,10 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
-<packages>
-  <package id="ClearScript.V8" version="5.4.6" targetFramework="net451" />
-  <package id="IronJS" version="0.2.0.1" targetFramework="net451" />
-  <package id="IronJS.Core" version="0.2.0.1" targetFramework="net451" />
-  <package id="Jint" version="2.9.1" targetFramework="net451" />
-  <package id="Jurassic" version="2.2.1" targetFramework="net451" />
-  <package id="NiL.JS" version="2.3.955.0" targetFramework="net451" />
-  <package id="Noesis.Javascript" version="0.7.1.0" targetFramework="net451" />
-</packages>
